♻️ api: explicit query params naming in DELETE jury comment route
