+++ api/tests/certification/session-management/acceptance/application/jury-comment-route_test.js
@@ -36,4 +36,27 @@ describe('Certification | Session Management | Acceptance | Application | Routes
       expect(response.statusCode).to.equal(204);
     });
   });
+
+  describe('DELETE /api/admin/sessions/{sessionId}/comment', function () {
+    it('should respond with 204', async function () {
+      // given
+      const server = await createServer();
+      await insertUserWithRoleSuperAdmin();
+      const session = databaseBuilder.factory.buildSession();
+      await databaseBuilder.commit();
+      const options = {
+        headers: {
+          authorization: generateValidRequestAuthorizationHeader(),
+        },
+        method: 'DELETE',
+        url: `/api/admin/sessions/${session.id}/comment`,
+      };
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(204);
+    });
+  });
 });
